+++ CHANGELOG.md
@@ -5,7 +5,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-### Documentation
+### Added
+- network-type private-bridge: add a new network layout, to provide private bridges for a group of pots
+- create-private-bridge: new subcommand to define and create a private-bridge
+- create: add option -B, to provide the bridge name if network-type is private-bridge
+- clone: add option -B, to provide the bridge name if network-type is private-bridge
+- prepare: add option -B, to provide the bridge name if network-type is private-bridge
+- destroy: add option -B, to provide a way to destroy a bridge
 - Image Guide: added a guide about how to create an Image of a pot
 
 ## [0.9.2] 2019-08-25